+++ packages/dds-components/src/components/Typography/Legend/Legend.stories.tsx
@@ -2,13 +2,13 @@ import preview from '#.storybook/preview';
 
 import { Legend } from '..';
 import { ddsProviderDecorator } from '../../../storybook';
-import { storyTypographyHtmlAttrs } from '../storyUtils';
+import { blockTgCommonArgTypes } from '../storyUtils';
 
 const meta = preview.meta({
   title: 'dds-components/Components/Typography/Legend',
   component: Legend,
   argTypes: {
-    ...storyTypographyHtmlAttrs,
+    ...blockTgCommonArgTypes,
   },
   decorators: [ddsProviderDecorator],
 });